+++ app/src/main/java/org/synapseworks/pageharbor/MainActivity.kt
@@ -1,0 +1,15 @@
+package org.synapseworks.pageharbor
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import org.synapseworks.pageharbor.ui.PageHarborApp
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            PageHarborApp()
+        }
+    }
+}